+++ cucumber-expressions/go/cucumber_expression_regexp_test.go
@@ -35,7 +35,7 @@ func TestCucumberExpressionRegExpTranslation(t *testing.T) {
 		assertRegexp(
 			t,
 			"I have {float} cukes at {int} o'clock",
-			`^I have (-?\d*\.\d+) cukes at ((?:-?\d+)|(?:\d+)) o'clock$`,
+			`^I have (-?\d*(?:[.,]\d+)?) cukes at ((?:-?\d+)|(?:\d+)) o'clock$`,
 		)
 	})