+++ cucumber-expressions/javascript/src/TreeRegexp.ts
@@ -38,7 +38,7 @@ export default class TreeRegexp {
         nonCapturingMaybe = false
       } else if (c === '?' && last === '(') {
         nonCapturingMaybe = true
-      } else if (c === ':' && nonCapturingMaybe) {
+      } else if ((c === ':' || c === '!') && nonCapturingMaybe) {
         stack[stack.length - 1].setNonCapturing()
         nonCapturingMaybe = false
       }