+++ cucumber-expressions/javascript/test/RegularExpressionTest.ts
@@ -32,11 +32,23 @@ describe('RegularExpression', () => {
   })
 
   it('transforms float without integer part', () => {
-    assert.deepStrictEqual(match(/(-?\d*\.?\d+)/, '.22')[0], 0.22)
+    assert.deepStrictEqual(
+      match(
+        new RegExp(`(${ParameterTypeRegistry.FLOAT_REGEXP.source})`),
+        '.22'
+      )[0],
+      0.22
+    )
   })
 
   it('transforms float with sign', () => {
-    assert.deepStrictEqual(match(/(-?\d*\.?\d+)/, '-1.22')[0], -1.22)
+    assert.deepStrictEqual(
+      match(
+        new RegExp(`(${ParameterTypeRegistry.FLOAT_REGEXP.source})`),
+        '-1.22'
+      )[0],
+      -1.22
+    )
   })
 
   it('returns null when there is no match', () => {