+++ gherkin/go/cmd/main.go
@@ -10,7 +10,7 @@ import (
 	b64 "encoding/base64"
 	"flag"
 	"fmt"
-	"github.com/cucumber/gherkin-go"
+	"github.com/cucumber/cucumber/gherkin/go"
 	"os"
 )
 
@@ -23,7 +23,7 @@ var dialectsFlag = flag.Bool("dialects", false, "print dialects as JSON")
 var defaultDialectFlag = flag.String("default-dialect", "en", "the default dialect")
 
 // Set during build with -ldflags
-var version string
+var version string = "(unknown version)"
 var gherkinDialects string
 
 func main() {